Modify function dispatching cross and per-thread data
Related-To: NEO-4585

+++ shared/test/unit_test/fixtures/command_container_fixture.h
@@ -46,9 +46,9 @@ struct WalkerThreadFixture {
     }
     void TearDown() {}
 
-    size_t startWorkGroup[3];
-    size_t numWorkGroups[3];
-    size_t workGroupSizes[3];
+    uint32_t startWorkGroup[3];
+    uint32_t numWorkGroups[3];
+    uint32_t workGroupSizes[3];
     uint32_t simd;
     uint32_t localIdDimensions;
     uint32_t requiredWorkGroupOrder;
